+++ SpeakHere/AudioViews/LevelMeter.m
@@ -176,7 +176,7 @@ int _cmp_levelThresholds(const void * a, const void * b)
 		if (_peakLevel > 0.)
 		{
 			peakLight = _peakLevel * _numLights;
-			if (peakLight >= _numLights) peakLight = _numLights - 1;
+			if (peakLight >= _numLights) peakLight = (int)(_numLights - 1);
 		}
 		
 		for (light_i=0; light_i<_numLights; light_i++)
